+++ after_deploy.sh
@@ -1,5 +1,5 @@
 npm install jsdoc
-git clone -b gh-pages "https://stellar-jenkins@github.com/stellar/js-stellar-base.git" jsdoc
+git clone -b gh-pages "https://payshares-jenkins@github.com/payshares/js-payshares-base.git" jsdoc
 
 if [ ! -d "jsdoc" ]; then
   echo "Error cloning"