+++ examples/signing-data.js
@@ -1,12 +1,12 @@
-import * as StellarBase from "../src"
+import * as PaysharesBase from "../src"
 
-var keypair = StellarBase.Keypair.random();
+var keypair = PaysharesBase.Keypair.random();
 var data = 'data to sign';
-var signature = StellarBase.sign(data, keypair.rawSecretKey());
+var signature = PaysharesBase.sign(data, keypair.rawSecretKey());
 
 console.log('Signature: '+signature.toString('hex'));
 
-if (StellarBase.verify(data, signature, keypair.rawPublicKey())) {
+if (PaysharesBase.verify(data, signature, keypair.rawPublicKey())) {
   console.log('OK!');
 } else {
   console.log('Bad signature!');