+++ src/account.js
@@ -1,22 +1,22 @@
 import BigNumber from 'bignumber.js';
 import isString from 'lodash/isString';
 import {Keypair} from './keypair';
-import {StrKey} from "./strkey";
+import {PsrKey} from "./psrkey";
 
 /**
  * Create a new Account object.
  *
- * `Account` represents a single account in Stellar network and its sequence number.
+ * `Account` represents a single account in Payshares network and its sequence number.
  * Account tracks the sequence number as it is used by {@link TransactionBuilder}.
- * See [Accounts](https://stellar.org/developers/learn/concepts/accounts.html) for more information about how
- * accounts work in Stellar.
+ * See [Accounts](https://payshares.org/developers/learn/concepts/accounts.html) for more information about how
+ * accounts work in Payshares.
  * @constructor
  * @param {string} accountId ID of the account (ex. `GB3KJPLFUYN5VL6R3GU3EGCGVCKFDSD7BEDX42HWG5BWFKB3KQGJJRMA`)
  * @param {string} sequence current sequence number of the account
  */
 export class Account {
   constructor(accountId, sequence) {
-    if (!StrKey.isValidEd25519PublicKey(accountId)) {
+    if (!PsrKey.isValidEd25519PublicKey(accountId)) {
       throw new Error('accountId is invalid');
     }
     if (!isString(sequence)) {
@@ -27,7 +27,7 @@ export class Account {
   }
 
   /**
-   * Returns Stellar account ID, ex. `GB3KJPLFUYN5VL6R3GU3EGCGVCKFDSD7BEDX42HWG5BWFKB3KQGJJRMA`
+   * Returns Payshares account ID, ex. `GB3KJPLFUYN5VL6R3GU3EGCGVCKFDSD7BEDX42HWG5BWFKB3KQGJJRMA`
    * @returns {string}
    */
   accountId() {