+++ src/network.js
@@ -2,12 +2,12 @@ import {hash} from "./hashing";
 
 /**
  * Contains passphrases for common networks:
- * * `Networks.PUBLIC`: `Public Global Stellar Network ; September 2015`
+ * * `Networks.PUBLIC`: `Public Global Payshares Network ; September 2015`
  * * `Networks.TESTNET`: `Test SDF Network ; September 2015`
  * @type {{PUBLIC: string, TESTNET: string}}
  */
 export const Networks = {
-	PUBLIC: "Public Global Stellar Network ; September 2015",
+	PUBLIC: "Payshares Public Global Network ; September 2017",
 	TESTNET: "Test SDF Network ; September 2015"
 };
 
@@ -15,7 +15,7 @@ var current = null;
 
 /**
  * The Network class provides helper methods to get the passphrase or id for different
- * stellar networks.  It also provides the {@link Network.current} class method that returns the network
+ * payshares networks.  It also provides the {@link Network.current} class method that returns the network
  * that will be used by this process for the purposes of generating signatures.
  *
  * You should select network your app will use before adding the first signature. You can use the `use`,
@@ -31,7 +31,7 @@ export class Network {
 	}
 
 	/**
-	 * Use Stellar Public Network
+	 * Use Payshares Public Network
 	 */
 	static usePublicNetwork() {
 		this.use(new Network(Networks.PUBLIC));