+++ test/test-helper.js
@@ -1,6 +1,6 @@
 if (typeof window === 'undefined') {
   require('babel/register');
-  global.StellarBase = require('../src/index');
+  global.PaysharesBase = require('../src/index');
   global.chai = require('chai');
   global.sinon = require('sinon');
   global.expect = global.chai.expect;